requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
fastapi-sso==0.10.0
python-multipart==0.0.6